Feat(Task 2-7 & 13): Code was updated functions.c

+++ functions.c
@@ -24,15 +24,18 @@ general * go_bypass(general *go, char *buffer, int size)
 		go->bol = functions_bin(go);
 	if (go->bol == 0)
 	{
+		go->res = (go->res == 0) ? 127 : 2;
 		if (isatty(STDIN_FILENO))
 			printf("%s: command not found\n", go->token[0]);
 		else
-			printf("%s: %d: %s: not found\n", go->exe, go->n, go->token[0]);
-		go->res = 127;
+		{
+			if (go->res == 2)
+				printf("%s: %d: '%s %s %s'\n", go->exe, go->n, go->token[0], go->token[1], go->token[2]);
+			else
+				printf("%s: %d: %s: not found\n", go->exe, go->n, go->token[0]);
+		}
 	}
-	printf("respuesta => %d\n", go->bol);
 	free(go->token);
-	
 	return (go);
 }
 
@@ -69,25 +72,66 @@ char *_access(char *token, envi *env)
 	return (NULL);
 }
 
+int validate_stream(general *go)
+{
+	int i = 0, res = 0;
+
+	while (go->token[i])
+	{
+		res = stream_match(go->token[i], go->token[i + 1], go->exe);
+		if (res == 1)
+			return (i);
+		if (res == -1)
+		{
+			go->res = 2;
+			return (-1);
+		}
+		if (res == 3)
+		{
+			go->fd = res;
+			return (i);
+		}
+		i++;
+	}
+	return (0);
+}
+
 int functions_bin(general *go)
 {
 	pid_t child = 0;
-	int status = 0;
+	int status = 0, res = -1;
 	char *path = NULL;
+	char **token = NULL;
 
-	printf("AAAAAAAAAAA\n");
 	if ((path = _access(go->token[0], go->env)) == NULL)
 		return (0);
+
 	child = fork();
 	if (child == 0)
 	{
-		execve(path, go->token, environ);
+		res = validate_stream(go);
+		if (res == -1)
+		{
+			go->res = 2;
+			return (0);
+		}
+		
+		if (res == 0)
+			execve(path, go->token, environ);
+		else
+			execve(path, _strncpy2(token, go->token, res), environ);
+		
 		kill(getpid(), SIGKILL);
 	}
 	else if (child > 0)
 	{
+		/*write(go->fd, buffa, 10);*/
+		
 		wait(&status);
 	}
 	free(path);
+	if (token)
+		free(token);
+	go->res = 0;
 	return (1);
 }
